Add SERVICE_HOST_DIR enviroment variable
Some scripts may want to provide output which the user can use on
their host system.  An example is showing a development script.  In
order for that to happen, the script needs to know the real location
of the files on the host filesystem.  This has now been shared through
an environment variable.