Remove links from dev scripts
These links were used and advised before but database is now called
database in most stacks and we don't depend on explicit linking anymore
in many applications.  The documentation value of the feature has eroded
with its inconsistent use and the database is generally named database
so as to not have to create the links for documentation.  Better have
the dev scripts reflect that because they'll be copied as a starting
point.